requirements.txt
requests
requests-oauthlib
pandas
matplotlib
python-dotenv
bokeh

